renamings and adding more strict type checking to vectors and nodes
